PlannerGUI - Prevent placements being updated if the task is not in the correct state.
The normal process is as follows: start a phase operation and task, then set placement status.

However, when placing components, failures occur, this failures need to be recorded.
HOWEVER, the placements list is NOT the place to record this.  An 'inspection/review' task should be
created instead and that show allow the failures to be recorded, otherwise users would not be able
to track and manage failues and act upon them.