PlannerGUI - Open issues in a different egui_dock node. Fix panic when main surface is empty.
Noting that:
* the egui_dock naming is terrible.
* the egui_dock api is terrible.
* floating the project explorer and then closing and opening various tabs can result in tabs opening in the wrong group.
* this workaround actually took hours and was extremely frustrating and still doesn't work very well.

To re-create the 'main surface empty' issue, float the project explorer, close all other tabs, then open another tab.

This is a 'temporary' fix.  A proper solution needs to somehow create groups of tabs. the 'project explorer' should be in a group, the 'issues' should be in another group. everything else should currently be in another group.  Opening 'parts' or 'placements' should never append the tab tab to the 'project explorer' or 'issues' group.